@@ -1,4 +1,4 @@
-﻿using System;
+﻿﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -7,14 +7,14 @@ namespace Shr2.Models
 {
     public class Config
     {
-        public string StorageConnectionString { get; set; }
+        public string StorageConnectionString { get; set; } = string.Empty;
 
-        public string StorageProvider { get; set; }
+        public string StorageProvider { get; set; } = "AzTableStorage";
 
-        public string Domain { get; set; }
+        public string Domain { get; set; } = string.Empty;
 
         public bool EncodeWithPermissionKey { get; set; }
 
-        public string[] PermissionKeys { get; set; }
+        public string[] PermissionKeys { get; set; } = Array.Empty<string>();
     }
 }
