@@ -1,12 +1,15 @@
-﻿using System;
+﻿﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using System.ComponentModel.DataAnnotations;
 
 namespace Shr2.Models
 {
     public class GStyleRequest
     {
-        public string LongUrl { get; set; }
+        [Required]
+        [Url]
+        public string LongUrl { get; set; } = string.Empty;
     }
 }
